% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire-ign.R
\name{acquire_cadastre}
\alias{acquire_cadastre}
\title{Acquiert le parcellaire cadastral (IGN WFS, optionnel)}
\usage{
acquire_cadastre(
  aoi,
  crs = 2154,
  cache_dir = tempdir(),
  overwrite = FALSE,
  country = "FR"
)
}
\arguments{
\item{aoi}{Objet \code{sf}/\code{sfc} d'emprise, dans le CRS cible.}

\item{crs}{Code EPSG de sortie. Défaut 2154.}

\item{cache_dir}{Répertoire de cache.}

\item{overwrite}{Re-télécharger même si le cache existe. Défaut \code{FALSE}.}

\item{country}{Code pays ISO. Défaut \code{"FR"}.}
}
\value{
Un objet \code{sf} de polygones de parcelles.
}
\description{
Acquiert le parcellaire cadastral (IGN WFS, optionnel)
}
