% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire-ign.R
\name{acquire_mnt}
\alias{acquire_mnt}
\title{Acquiert le MNT depuis RGE ALTI (IGN WMS)}
\usage{
acquire_mnt(
  aoi,
  res_m = 5,
  crs = 2154,
  cache_dir = tempdir(),
  overwrite = FALSE,
  country = "FR"
)
}
\arguments{
\item{aoi}{Objet \code{sf}/\code{sfc} d'emprise, dans le CRS cible.}

\item{res_m}{Résolution du raster (m). Défaut 5.}

\item{crs}{Code EPSG de sortie. Défaut 2154.}

\item{cache_dir}{Répertoire de cache.}

\item{overwrite}{Re-télécharger même si le cache existe. Défaut \code{FALSE}.}

\item{country}{Code pays ISO. Défaut \code{"FR"}.}
}
\value{
Le chemin du raster \code{mnt.tif} écrit en cache.
}
\description{
Acquiert le MNT depuis RGE ALTI (IGN WMS)
}
