% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cable_calcul_xs}
\alias{cable_calcul_xs}
\title{Position horizontale du câble à l'abscisse curviligne \code{s}.}
\usage{
cable_calcul_xs(th, tv, lo, eao, w, f, s1, s)
}
\description{
Position horizontale du câble à l'abscisse curviligne \code{s}.
}
