% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cable_calcul_zs}
\alias{cable_calcul_zs}
\title{Position verticale (chute depuis le support haut) à l'abscisse curviligne
\code{s} : c'est elle qui fournit la garde au sol du câble.}
\usage{
cable_calcul_zs(th, tv, lo, eao, w, f, s1, s)
}
\description{
Position verticale (chute depuis le support haut) à l'abscisse curviligne
\code{s} : c'est elle qui fournit la garde au sol du câble.
}
