% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cable_f_x}
\alias{cable_f_x}
\title{Equation horizontale de la caténaire élastique \code{f_x(Th, Tv)}, nulle à la
solution.}
\usage{
cable_f_x(th, tv, lo, eao, w, f, s1, d)
}
\description{
Equation horizontale de la caténaire élastique \code{f_x(Th, Tv)}, nulle à la
solution.
}
