% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cable_f_z}
\alias{cable_f_z}
\title{Equation verticale de la caténaire élastique \code{f_z(Th, Tv)}, nulle à la
solution.}
\usage{
cable_f_z(th, tv, lo, eao, w, f, s1, h)
}
\description{
Equation verticale de la caténaire élastique \code{f_z(Th, Tv)}, nulle à la
solution.
}
