% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cable_find_thtv_tmax}
\alias{cable_find_thtv_tmax}
\title{Amorçage \verb{(Th, Tv, faisable)} par recherche sur grille sous la tension
admissible. Renvoie un vecteur de longueur 3 ; \code{faisable} vaut 1 si
\code{sqrt(Th^2 + Tv^2) <= tmax}, 0 sinon.}
\usage{
cable_find_thtv_tmax(tmax, w, eao, f, pas, d, h, lo, step)
}
\description{
Amorçage \verb{(Th, Tv, faisable)} par recherche sur grille sous la tension
admissible. Renvoie un vecteur de longueur 3 ; \code{faisable} vaut 1 si
\code{sqrt(Th^2 + Tv^2) <= tmax}, 0 sinon.
}
