% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cable_newton_thtv}
\alias{cable_newton_thtv}
\title{Résout \code{f_x = f_z = 0} (tensions \verb{Th, Tv} au support haut) par
Newton-Raphson à Jacobien analytique, repli sur grille. Renvoie un vecteur
\code{c(Th, Tv)}.}
\usage{
cable_newton_thtv(th, tv, h_alt, d, lo, w, s1, f, eao, tmax, err)
}
\description{
Résout \code{f_x = f_z = 0} (tensions \verb{Th, Tv} au support haut) par
Newton-Raphson à Jacobien analytique, repli sur grille. Renvoie un vecteur
\code{c(Th, Tv)}.
}
