% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasources.R
\name{get_national_crs}
\alias{get_national_crs}
\title{CRS national d'un pays}
\usage{
get_national_crs(country = "FR")
}
\arguments{
\item{country}{Code pays ISO. Défaut \code{"FR"}.}
}
\value{
Un entier (code EPSG ; 2154 pour la France).
}
\description{
CRS national d'un pays
}
\examples{
get_national_crs("FR")
}
