% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasources.R
\name{list_countries}
\alias{list_countries}
\title{Liste les pays configurés}
\usage{
list_countries()
}
\value{
Un vecteur de codes pays ISO.
}
\description{
Liste les pays configurés
}
\examples{
list_countries()
}
