% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-provenance.R
\name{politique_cache_valeurs}
\alias{politique_cache_valeurs}
\title{Valeurs admises pour \code{politique_cache}}
\usage{
politique_cache_valeurs()
}
\value{
Un vecteur de caractères.
}
\description{
Valeurs admises pour \code{politique_cache}
}
\seealso{
\code{\link[=cache_utilisable]{cache_utilisable()}}, \code{specs/027}.
}
