% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage-postgis.R
\name{sb_ensure_schema}
\alias{sb_ensure_schema}
\title{Crée le schéma PostgreSQL du backend s'il n'existe pas}
\usage{
sb_ensure_schema(backend)
}
\arguments{
\item{backend}{Objet \code{foretaccess_storage_postgis}.}
}
\value{
\code{backend} de façon invisible.
}
\description{
Crée le schéma PostgreSQL du backend s'il n'existe pas
}
