% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{sb_list_layers}
\alias{sb_list_layers}
\title{Liste les couches disponibles}
\usage{
sb_list_layers(backend, ...)
}
\arguments{
\item{backend}{Objet de stockage (\code{\link[=storage_gpkg]{storage_gpkg()}} ou \code{\link[=storage_postgis]{storage_postgis()}}).}

\item{...}{Arguments spécifiques au backend.}
}
\value{
Un vecteur de caractères (noms de couches).
}
\description{
Liste les couches disponibles
}
