% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{sb_read_layer}
\alias{sb_read_layer}
\title{Relit une couche vectorielle}
\usage{
sb_read_layer(backend, layer, ...)
}
\arguments{
\item{backend}{Objet de stockage (\code{\link[=storage_gpkg]{storage_gpkg()}} ou \code{\link[=storage_postgis]{storage_postgis()}}).}

\item{layer}{Nom de la couche (chaîne).}

\item{...}{Arguments spécifiques au backend.}
}
\value{
Un objet \code{sf}.
}
\description{
Relit une couche vectorielle
}
