% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{sb_write_layer}
\alias{sb_write_layer}
\title{Écrit une couche vectorielle}
\usage{
sb_write_layer(backend, layer, data, ...)
}
\arguments{
\item{backend}{Objet de stockage (\code{\link[=storage_gpkg]{storage_gpkg()}} ou \code{\link[=storage_postgis]{storage_postgis()}}).}

\item{layer}{Nom de la couche (chaîne).}

\item{data}{Objet \code{sf}.}

\item{...}{Arguments spécifiques au backend.}
}
\value{
\code{backend} de façon invisible.
}
\description{
Écrit une couche vectorielle
}
