% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_config}
\alias{write_config}
\title{Écrit une configuration au format YAML}
\usage{
write_config(cfg, path)
}
\arguments{
\item{cfg}{Objet \code{foretaccess_config}.}

\item{path}{Chemin de sortie.}
}
\value{
\code{path} de façon invisible.
}
\description{
Écrit une configuration au format YAML
}
