% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foretaccess-package.R
\docType{package}
\name{foretaccess-package}
\alias{foretaccess}
\alias{foretaccess-package}
\title{foretaccess: Cartographie de l'Accessibilité des Forêts (Réimplémentation de Sylvaccess)}
\description{
Cartographie automatique de l'accessibilité des forêts selon le mode d'exploitation (skidder, porteur, câble-mât, camion DFCI). Réimplémentation moderne, découplée et testable du modèle Sylvaccess (INRAE, S. Dupire), avec un cœur métier pur en R et un noyau câble en Rust (via 'extendr'), et des sorties en base spatiale (PostGIS / GeoPackage).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pobsteta/foretaccess}
  \item Report bugs at \url{https://github.com/pobsteta/foretaccess/issues}
}

}
\author{
\strong{Maintainer}: Pascal Obstetar \email{pascal.obstetar@gmail.com}

Authors:
\itemize{
  \item Pascal Obstetar \email{pascal.obstetar@gmail.com}
}

}
\keyword{internal}
